perf: use float variants for mathematical operations in calc function

+++ src/ratio_calculate.cpp
@@ -80,23 +80,23 @@ void calc(int x, int y)
 
   constexpr float center_x = 256;
   constexpr float center_y = 256;
-  constexpr float max_distance = 362.039;
+  constexpr float max_distance = 362.039f;
   constexpr float inv_max_distance = 1.0f / max_distance;
 
-  float dx = center_x - x - 0.5;
-  float dy = center_y - y - 0.5;
-  float distance = std::sqrt(square_sum(dx, dy));
+  float dx = center_x - x - 0.5f;
+  float dy = center_y - y - 0.5f;
+  float distance = std::sqrtf(square_sum(dx, dy));
 
   float distance_scale = distance * inv_max_distance;
   float eccentricity = 0.9 * distance_scale;
 
   float sigma_minor = 0.5f + distance_scale * 2.5f;
   float denom = 1.0f - eccentricity * eccentricity;
-  float sigma_major = sigma_minor / std::sqrt(denom);
+  float sigma_major = sigma_minor / std::sqrtf(denom);
 
-  float angle = std::atan2(dy, dx);
-  float cos_angle = std::cos(angle);
-  float sin_angle = std::sin(angle);
+  float angle = std::atan2f(dy, dx);
+  float cos_angle = std::cosf(angle);
+  float sin_angle = std::sinf(angle);
 
   float normalization_factor = 1.0f / (2.0f * static_cast<float>(M_PI) * sigma_major * sigma_minor);
 
@@ -114,7 +114,7 @@ void calc(int x, int y)
         float major_coord = (delta_x * cos_angle + delta_y * sin_angle) / sigma_major;
         float minor_coord = (delta_x * sin_angle - delta_y * cos_angle) / sigma_minor;
         float exponent = (square_sum(major_coord, minor_coord)) * 0.5f;
-        float psf_value = normalization_factor * std::exp(-exponent);
+        float psf_value = normalization_factor * std::expf(-exponent);
 
         psf_s[psf_len] = psf_value;
         psf_c[psf_len] = c;
